social_media_api/requirements.txt
# Core
Django==4.2.0
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0
djangorestframework-authtoken==1.4.0

# Database
mysqlclient==2.2.4
psycopg2-binary==2.9.9  # For PostgreSQL

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# Security
django-cors-headers==4.0.0

# Media Files
Pillow==10.0.0

# Production Server
waitress==2.1.2  # For Windows
gunicorn==21.2.0  # For Linux

# Documentation
drf-yasg==1.21.7

# Development (optional)
django-debug-toolbar==4.2.0
django-extensions==3.2.3